Remove watchdog thread & fix socket checking
For some as-of-yet unknown reason, sending empty
batches to STINGER in a separate thread messes
up the connection to the point that further
legitimate batches don't get saved.

Thus, the watchdog thread which uses empty
batches to check the socket ultimately
prevents the real batches from being saved.

This commit:

1. Removes the watchdog thread

2. Adds a new verify_connection() method to
   StingerStream which will check the socket
   handle and attempt reconnection 5 times
   before finally raising a RuntimeError

3. Catches socket write exceptions during
   batch sending and invokes verify_connection()

The result is decent socket hardening that
won't immediately catch problems, but will
reliably either reconnect or raise an
exception up to the original caller.

Thus, if you require reliable writes,
as long as you pass immediate=true to the
insert/ endpoint, you are guaranteed
either a 201 Created or 500 error,
(whereas previously an insertion failure
could still result in a 201)